+++ NEWS.md
@@ -5,6 +5,7 @@ version 0.1.4
 =============
 
 * Cache arg `sideEffect` can now be a path
+* Cache arg `digestPathContent` default changed from FALSE (was for speed) to TRUE (for content accuracy)
 * New function, `searchFull`, which shows the full search path, known alternatively as "scope", or "binding environments". It is where R will search for a function when requested by a user.
 * Uses memoise::memoise for several functions (`loadFromLocalRepo`, `pkgDep`, `package_dependencies`, `available.packages`) for speed -- will impact memory at the expense of speed
 * New `Require` function that attempts to create a lighter weight package reproducibility chain. This function is usable in a reproducible workflow: it includes both installing and loading of packages, it maintains version numbers, and uses smart caching for speed. In tests, it can evaluate whether 20 packages and their dependencies (~130 packages) are installed and loaded in ~0.1 seconds. This is much slower than running require on those 20 packages, but this may be fast enough for many purposes.